+++ src/loader/parameters.cpp
@@ -46,7 +46,7 @@ void Parameters::load_config(nlohmann::json& header){
 
 void Parameters::load_tensor(std::unordered_map<std::string, std::unique_ptr<Tensor>>& m, char* p, const std::string& key, nlohmann::json& value){
     uint64_t offset = value["offset"];
-    std::vector<int64_t> shape = value["shape"];
+    std::vector<size_t> shape = value["shape"];
 
     std::unique_ptr<Tensor> t = std::make_unique<Tensor>(reinterpret_cast<float*>(p + offset), shape);
     m.insert({key, std::move(t)});
